+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Fluent
 
+## Unreleased
+
+### Fixed
+
+- SBCL: Occasional issues involving `simple-base-string` in path components.
+
 ## 1.0.0 (2025-07-31)
 
 Initial release.